Rename containers folder
