+++ app/src/components/Contacts/Contacts.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import { Link } from 'react-router';
 
 export default class Contacts extends React.Component {
   render() {
